[feat] WeeklyTheme textcolor 변경
