+++ screens/HomeScreen/HomeScreen.tsx
@@ -8,7 +8,7 @@ import { ScrollView, View } from "react-native";
 // components
 import Input from "../../components/Input/Input";
 import HomeTitle from "./components/HomeTitle/HomeTitle";
-import WeeklyTheme from "./components/WeeklyTheme/WeeklyThemeTest";
+import WeeklyTheme from "./components/WeeklyTheme/WeeklyTheme";
 import NearByTheme from "./components/NearByTheme/NearByTheme";
 import BookmarkedTheme from "./components/BookmarkedTheme/BookmarkedTheme";
 import Header from "../../components/Header/Header";
